added comment deletion and creating

+++ src/main/java/com/example/microgram/DAO/PostDao.java
@@ -1,25 +1,19 @@
 package com.example.microgram.DAO;
 
 import com.example.microgram.DTO.PostDto;
-import com.example.microgram.DTO.PostUserDto;
 import com.example.microgram.DTO.PostUserImageDto;
 import com.example.microgram.Utility.DataGenerator.PostExample;
 import com.example.microgram.Utility.FileUtil;
-import com.example.microgram.Utility.Utils;
 import lombok.RequiredArgsConstructor;
-import org.springframework.core.io.ByteArrayResource;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Component;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStream;
 import java.sql.PreparedStatement;
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @Component
 @RequiredArgsConstructor
@@ -94,7 +88,7 @@ public class PostDao {
         jdbcTemplate.update(query, fileName, post.getDescription(), Timestamp.valueOf(ld), getUserIdByUsername(post.getUsername()));
 
         return PostDto.builder()
-                .id(getUserIdByUsername(post.getUsername()))
+                .id(getIdByImage(fileName))
                 .image(fileName)
                 .description(post.getDescription())
                 .time(ld)
@@ -106,6 +100,11 @@ public class PostDao {
         if (!isPostOwner(postID, username))
             return "Вы не можете удалять чужие публикации!";
         deleteCommentsOfPost(postID);
+        deleteLikesOfPost(postID);
+        var optImage = getImageById(postID);
+        if (optImage.isEmpty())
+            return "Не существует такой публикации!";
+        FileUtil.deleteImageFile(optImage.get());
         String query = "DELETE FROM POSTS\n" +
                 "WHERE id = ? AND user_id = ?";
         jdbcTemplate.update(query, postID, getUserIdByUsername(username));
@@ -117,11 +116,22 @@ public class PostDao {
         return list.size();
     }
 
-    private boolean isPostOwner(Long postID, String username) {
+    public boolean isPostOwner(Long postID, String username) {
         String query = "SELECT COUNT(id) FROM POSTS\n" +
                 "WHERE id = ? AND user_id = ?;";
         var result = jdbcTemplate.queryForObject(query, Integer.class, postID, getUserIdByUsername(username));
         return result == 1;
+    }
+
+    private Optional<String> getImageById(Long postID) {
+        String query = "SELECT image FROM POSTS WHERE id = ?";
+        String image = jdbcTemplate.queryForObject(query, String.class, postID);
+        return Optional.of(image);
+    }
+
+    private Long getIdByImage(String image) {
+        String query = "SELECT id FROM posts WHERE image = ?";
+        return jdbcTemplate.queryForObject(query, Long.class, image);
     }
 
     private Long getUserIdByUsername(String username) {
